systemX/Price_Generator/requirements.txt
matplotlib
numpy
sklearn
tensorflow
pandas
pandas_datareader
mysql-connector-python
yfinance
python-dotenv
